Moved types for both workers to same place
